python27: support test/support
In the maintenance release bump in
90059701a836b0d0dbc676e2d1ce417cc5a25bd2 a certain change to /test/ was
backported from Python 3:

- bpo-30207: To simplify backports from Python 3, the test.test_support
  module was converted into a package and renamed to test.support.  The
  test.script_helper module was moved into the test.support package.
  Names test.test_support and test.script_helper are left as aliases to
  test.support and test.support.script_helper.

+++ pkgs/development/interpreters/python/cpython/2.7/default.nix
@@ -163,7 +163,8 @@ in stdenv.mkDerivation {
         # functionality to 2.x from 3.x
         for item in $out/lib/python${majorVersion}/test/*; do
           if [[ "$item" != */test_support.py*
-             && "$item" != */regrtest.py* ]]; then
+             && "$item" != */test/support
+             && "$item" != */test/regrtest.py* ]]; then
             rm -rf "$item"
           else
             echo $item
